separate spawnable types into file finish commenting store files
