+++ thetawave_lib/src/spawnable/mod.rs
@@ -2,4 +2,9 @@
 
 pub mod components;
 pub mod resources;
+mod spawnable_types;
 pub mod systems;
+
+pub use self::spawnable_types::{
+    AllyType, ConsumableType, EffectType, EnemyType, ItemType, MobType, NeutralType, SpawnableType,
+};
