+++ thetawave_lib/src/spawnable/resources/mod.rs
@@ -3,6 +3,7 @@ use crate::{
     motion::components::{Hitbox2DComponent, Motion2DComponent},
     spawn::components::DespawnAtBorderComponent,
     spawnable::components::BlastComponent,
+    spawnable::SpawnableType,
     visual::resources::SpriteSheetsResource,
 };
 
@@ -11,8 +12,6 @@ use amethyst::{
     ecs::prelude::{Builder, Entities, LazyUpdate, ReadExpect},
     renderer::{SpriteRender, Transparent},
 };
-
-use serde::{Deserialize, Serialize};
 
 mod modifiers;
 mod spawnables;
@@ -25,96 +24,6 @@ pub use self::{
         ThrusterEntityData,
     },
 };
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum SpawnableType {
-    Consumable(ConsumableType),
-    Item(ItemType),
-    Effect(EffectType),
-    Mob(MobType),
-}
-
-impl SpawnableType {
-    pub fn get_spritesheet_name(&self) -> String {
-        match self {
-            &Self::Consumable(_) => String::from("consumables"),
-            &Self::Effect(_) => String::from("effects"),
-            &Self::Item(_) => String::from("items"),
-            &Self::Mob(_) => String::from("mobs"),
-        }
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum MobType {
-    Enemy(EnemyType),
-    Ally(AllyType),
-    Neutral(NeutralType),
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum EnemyType {
-    Pawn,
-    Drone,
-    StraferRight,
-    StraferLeft,
-    MissileLauncher,
-    Missile,
-    RepeaterBody,
-    RepeaterHead,
-    RepeaterLeftShoulder,
-    RepeaterRightShoulder,
-    RepeaterLeftArm,
-    RepeaterRightArm,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum AllyType {
-    Hauler,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum NeutralType {
-    MoneyAsteroid,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum ConsumableType {
-    DefenseWrench,
-    Money1,
-    Money5,
-    HealthWrench,
-    Armor,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum ItemType {
-    SteelBarrel,
-    PlasmaBlasts,
-    HazardousReactor,
-    WarpThruster,
-    Tentaclover,
-    DefenseSatellite,
-    DoubleBarrel,
-    YithianPlague,
-    Spice,
-    EnhancedPlating,
-    StructureReinforcement,
-    BlasterSizeEnhancer,
-    FrequencyAugmentor,
-    TractorBeam,
-    BlastRepeller,
-}
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum EffectType {
-    AllyBlastExplosion,
-    EnemyBlastExplosion,
-    PoisonBlastExplosion,
-    CriticalBlastExplosion,
-    MobExplosion,
-    Star, //TODO: implement background stars
-    Giblets(MobType),
-}
 
 /// Spawn any kind of spawnable entity
 pub fn spawn_spawnable(
