+++ thetawave_lib/src/tools/random.rs
@@ -1,5 +1,7 @@
 use rand::{thread_rng, Rng};
 
+/// Randomly picks index of vector using weights
+/// Takes in a vector of weights
 pub fn weighted_rng(probs: Vec<f32>) -> usize {
     let prob_space = probs.iter().fold(0.0, |sum, prob| sum + prob);
     let pos = thread_rng().gen::<f32>() * prob_space;
